pyproject.toml
[project]
name = "ddf-ai-enrichment-pipeline"
version = "0.1.1"
description = "AI enrichment pipeline for Direct Democracy Forum"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.54.0",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.0",
]

[dependency-groups]
dev = [
    "ipython>=8.32.0",
    "ruff>=0.9.7",
]
